Address review comments on Scottish Child Payment
Changes:
- Update SCP amount to £27.15 for April 2025
- Split variables into separate files (is_scp_eligible_child.py)
- Add age-specific takeup rates: 97% for under-6s, 86% for 6+
- Add comment about CTC/WTC phase-out in qualifying benefits
