+++ policyengine_uk/parameters/gov/social_security_scotland/scottish_child_payment/amount.yaml
@@ -5,6 +5,7 @@ values:
   2022-04-01: 20
   2022-11-14: 25
   2024-04-01: 26.70
+  2025-04-01: 27.15
 metadata:
   unit: currency-GBP
   period: week
