--- policyengine_uk/parameters/gov/social_security_scotland/scottish_child_payment/takeup_rate.yaml
@@ -1,13 +1,0 @@
-description: >
-  Take-up rate for Scottish Child Payment. Estimated based on comparison
-  of recipients to eligible population. SCP has relatively high take-up
-  due to automatic awards for some recipients and targeted outreach.
-values:
-  0001-01-01: 0.85
-metadata:
-  unit: /1
-  period: year
-  label: Scottish Child Payment take-up rate
-  reference:
-    - title: SPICe analysis of SCP coverage
-      href: https://spice-spotlight.scot/2024/07/31/scottish-child-payment-who-gets-it/
